ADD:  multilanguage support   (english and swedish)
To add a new language,   add a text file to the folder Language with the name of Language as filename
ie:
Language\Deutsch.txt    for a german translator.

This supports only the control.Text property. It will need to expand to the hexboxes contextmenus aswell.

Important!
Save the file in UTF-8 in order to get localized characters.